backend/requirements.txt
# NavAI Real-Time Backend Dependencies
# FastAPI + WebSocket + AI Services

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
websockets==12.0
httpx==0.27.0

# AI Services
groq==0.4.2
openai==1.12.0

# Audio/TTS
edge-tts==6.1.9

# Image Processing
pillow==10.2.0
numpy==1.26.3

# Utilities
pydantic==2.6.0
python-multipart==0.0.9

# Optional: Report Generation
reportlab==4.0.9
